improve Box component
