+++ docs/src/components/Box/styles.module.css
@@ -1,0 +1,19 @@
+.container {
+  display: flex;
+  align-items: center;
+  padding: 1em;
+  border-radius: 8px;
+  cursor: pointer;
+  transition: all 0.5s ease-in-out;
+}
+
+.container:hover {
+  box-shadow: 0px 0px 0px 4px #4bb3fd;
+}
+
+svg {
+  color: white;
+  height: 32px;
+  width: 32px;
+  margin-right: 12px;
+}
